+++ probe-rs-tools/Cargo.toml
@@ -18,12 +18,11 @@ default-run = "probe-rs"
 [features]
 default = []
 # Include server/client functionality
-remote = ["dep:tokio-tungstenite", "dep:axum", "probe-rs-rpc/remote", "probe-rs-rpc/axum", "probe-rs-rpc-client/remote"]
+remote = ["dep:axum", "probe-rs-rpc/remote", "probe-rs-rpc-client/remote"]
 
 [dependencies]
 anyhow = { workspace = true }
 base64 = "0.23"
-docsplay = { workspace = true }
 jep106 = "0.3"
 num-traits = "0.2"
 scroll = "0.13"
@@ -113,12 +112,8 @@ probe-rs-rpc-client = { workspace = true }
 tokio = { version = "1.0", features = ["full"] }
 tokio-util = { version = "0.7", features = ["codec"] }
 futures-util = { version = "0.3" }
-tokio-tungstenite = { version = "0.30", features = [
-    "rustls-tls-webpki-roots",
-], optional = true }
 axum = { version = "0.8", features = ["ws"], optional = true }
 tempfile = "3.0"
-postcard = { version = "1.0", features = ["use-std"] }
 postcard-rpc = { version = "0.12.0", features = ["use-std"] }
 postcard-schema = { version = "0.2.0", features = ["use-std", "derive"] }
 sha2 = "0.11"
